Spotlights stores angles, adding comment
Note that FOVs are expressed as HALVES of their full angle.
Passing a Camera constructor a fov of 45 degrees means having a 90
degree field of view. In the same way, saying the outer cutoff angle of
the spotlight is 60 degrees corresponds to having a cone of light whose
top angle is 120 degrees.