Scaling the scene in the range [0, 3]
Camera needs to be centered.
It is now clear that we have a problem with the intensity of the
indirect light. However, the problem might be solved with the
normalization step that has been missing.

Currently, the most reasonable normalization step we found in other
implementations is multiplying by 4*PI/N, where N is the number of
samples.