Implementation for interpolation
TODO: extract the lambda in a specific struct - maybe Interpolator is a
good name? maybe just move the logic in BlockInfo? - so that the code is
cleaner.

TODO: move computing points after interpolation in another function,
giving it the number of points on the edges, the interpolator structure
and a reference to the block we're filling.