Fixed the problem of the previous commit
The problem was found when interpolating on the last column or row,
since the computation of the points to be added was changing.
However, the step size to be used had to stay the same.

TODO: refactor those loops so that naming is clearer, results are easily
obtained, cartesian_product can be explicited and then a simple
transform can be used to get a block, applying the interpolator.