Setup for direct shading pt. 2, completed
A directly lit scene can be visualized here.
In the deferred direct pass, we render the frame to a texture which
we'll show later through rendering on a quad.

Next step is creating the shadow maps and integrating those in the
direct pass. The following step will be rendering the IES masks so that
a reflective shadow map can be used instead of a simple shadow map.