Refactoring: extracting methods, setup for events
Extracted methods to make code more readable.

Set up an on_camera_moved method, to be used when an event signals the
camera has been moved and the buffers using camera informations need to
be updated accordingly.

Note: this is to be kept like this until the event system is reworked or
the CameraMoved event is introduced, so that we can benchmark the
differences.