IES Masking refactoring: method extraction
Extracted utility function for creating VAO from IES file.
Extracted methods for IES Mask structures' initialization and usage.
Extracted function for computing the model matrix for a light source.

NOTE: the abovementioned functions might be refactored and generalized
so that they're not methods of DeferredLayer class; in fact, they do not
belong there and they might be methods we write once and reuse in many
layers. Consider this.