Adapting SceneLayer to event system
Note that SceneLayer doesn't currently push any event to the event
queue.

Some parameters where fixed to commonly used values for the comparisons
that we will be doing. In particular, drawing the wireframe of the IES
light was disabled and the displacement radius when sampling the texture
was brought to 2.0f.