Setup for loading textures from the scene
Note that this is not working and will result in a nullptr being
returned by the stbi_load function. This is normal, since the
string_view we pass in is incorrect and needs to be computed better.

Also note that the #define added in main.cpp is needed because of this:
	https://gamedev.stackexchange.com/a/158107/111281