Loading the textures in the correct way
Added a field to the SceneLoader class, so that it can store the parent
directory of the scene file as a std::filesystem::path.
Added initialization of the above-mentioned path in the constructor.
When the path to the texture is extracted, it is appended to the
directory path and it is used to load the texture.