Setup for resource management
This is needed to avoid loading textures that are shared between objects
more than once. In a way, it is like keeping a map whose keys are the
hashes of the textures' paths while the values are the pointers to the
loaded texture; however, we delete the hashes once we are done with
loading the textures. The connection between objects and the vector of
(pointers to) resources is done through an index stored in the
SceneObject, pointing at the right position in the vector of textures.