Loading textures to GPU memory working correctly
Note that the loading process works correctly since it doesn't throw any
exception or any other kind of error.
Also note that a fix was necessary to make it work: the loading function
pushed only a fourth of the needed bytes, and this resulted in trying to
read out of bounds.