Update to shaders for multiple lights
Updated deferred_direct and deferred_indirect fragment shaders so that
the computation supports using more than one light in the scene.
Note that this is yet to be tested.

Changed the number of bits used to represent normals and fluxes in the
reflective shadow maps, from 16 bits per channel to 8 bits per channel.