Refining wireframe drawing, pt.1
Drawing wireframe on another framebuffer, with a separate texture. This
texture has to be conditionally composed (TODO) in the last pass, after
direct and indirect lighting have been drawn to the screen quad.

Note that the depth buffer used when filling the g-buffers is used for
this pass, in order to retain the depth information.

TODO: refactor the wireframe drawing function out of the update
function.